+++ ccptools/dtu/casting/__init__.py
@@ -1,3 +1,5 @@
 from ._filetime import *
 from ._string import *
+from ._timestamp import *
+from ._instant import *
 from ._any import *